@jeremyong-az 2021-07-22
NOTE: don't check in until a followup commit addressing compile errors due to downstream headers/source files relying on `binary_semaphore.h` (which was removed)