@jeremyong-az 2021-07-22
> Looks great!
> 
> Since we control AZStd::binary_semaphore, and I very much doubt any usage requires or depends on IPC, maybe we should replace the guts of AZStd::binary_semaphore with a condition variable like you've done here?

Yea I agree didn't want to do such an invasive change here just yet (and it would change the size of the binary_semaphore) but we should definitely investigate that option