@jeremyong-az 2021-07-23
Ah, I messed up, I should have looked at the implementation of our condition_variable more carefully. It doesn't spin to check for spurious wakeups on behalf of the user.