@jeremyong-az 2021-07-23
Closing! In favor of a different PR I will put up in the near future :)

So, interestingly, my intuition about Win32 Events is slightly out-of-date. Once I account for spurious wakeups (user error, the condition variable class I used in a previous codebase did this internally), the speedup is present but small enough that it could be attributable to noise. As an experiment, I reimplemented `binary_semaphore` to the following:

```c++
    class binary_semaphore
    {
    public:
        binary_semaphore(bool initialState = false)
            : m_state{ initialState }
        {
        }

        binary_semaphore(const char* name, bool initialState = false)
            : m_state{ initialState }
        {
            (void)name;
        }

        binary_semaphore(const binary_semaphore&) = delete;
        binary_semaphore& operator=(const binary_semaphore&) = delete;

        void acquire()
        {
            while (true)
            {
                bool result = m_state.load(memory_order_relaxed);
                while (!m_state)
                {
                    WaitOnAddress(&m_state, &result, sizeof(m_state), AZ_INFINITE);
                    result = true;
                }

                if (m_state.compare_exchange_strong(result, false))
                {
                    return;
                }
            }
        }

        void release()
        {
            m_state.store(true, memory_order_release);
            WakeByAddressSingle(&m_state);
        }

private:
        atomic<bool> m_state;
};
```

Even with this lean implementation, the improvement I'm seeing is nominal (if these benchmarks are to be believed). The issue with our `condition_variable` implementation which I believe we should investigate in the future is the use of a normal mutex instead of an SRW mutex (slim reader writer) [link](https://docs.microsoft.com/en-us/windows/win32/sync/slim-reader-writer--srw--locks). Interestingly, I decided to verify this implementation against the C++20 atomic wait implementation in the STL, and surprisingly, they spin! For reference: https://github.com/microsoft/STL/blob/9a9820df1a1d3fa84100e3169ff37fdd4fa41759/stl/inc/atomic#L580 

In other words, I apparently have no idea what I'm talking about, at least with more recent versions of the kernel so I'm going to look forward to the more important optimization of relaxing prioritization granularity instead.